outsource some genericish methods
